+++ init.php
@@ -28,6 +28,7 @@ JLoader::registerNamespace('OAuth2', PRISM_PATH_LIBRARY . '/libs');
 JLoader::registerNamespace('RandomLib', PRISM_PATH_LIBRARY . '/libs');
 JLoader::registerNamespace('SecurityLib', PRISM_PATH_LIBRARY . '/libs');
 JLoader::registerNamespace('AdamPaterson', PRISM_PATH_LIBRARY . '/libs');
+JLoader::registerNamespace('Carbon', PRISM_PATH_LIBRARY . '/libs');
 
 // Register some helpers.
 JHtml::addIncludePath(PRISM_PATH_LIBRARY .'/ui/helpers');