+++ init.php
@@ -17,18 +17,21 @@ jimport('joomla.filesystem.file');
 jimport('joomla.filesystem.folder');
 jimport('joomla.filesystem.path');
 
+JLoader::registerNamespace('Abraham', PRISM_PATH_LIBRARY . '/libs');
+JLoader::registerNamespace('AdamPaterson', PRISM_PATH_LIBRARY . '/libs');
 JLoader::registerNamespace('Aws', PRISM_PATH_LIBRARY . '/libs');
+JLoader::registerNamespace('Carbon', PRISM_PATH_LIBRARY . '/libs');
 JLoader::registerNamespace('Defuse', PRISM_PATH_LIBRARY . '/libs');
-JLoader::registerNamespace('Prism', JPATH_LIBRARIES);
-JLoader::registerNamespace('League', PRISM_PATH_LIBRARY . '/libs');
-JLoader::registerNamespace('Psr', PRISM_PATH_LIBRARY . '/libs');
+JLoader::registerNamespace('Facebook', PRISM_PATH_LIBRARY . '/libs');
+JLoader::registerNamespace('Google', PRISM_PATH_LIBRARY . '/libs');
 JLoader::registerNamespace('GuzzleHttp', PRISM_PATH_LIBRARY . '/libs');
 JLoader::registerNamespace('JmesPath', PRISM_PATH_LIBRARY . '/libs');
-JLoader::registerNamespace('OAuth2', PRISM_PATH_LIBRARY . '/libs');
+JLoader::registerNamespace('League', PRISM_PATH_LIBRARY . '/libs');
+JLoader::registerNamespace('Monolog', PRISM_PATH_LIBRARY . '/libs');
+JLoader::registerNamespace('Prism', JPATH_LIBRARIES);
+JLoader::registerNamespace('Psr', PRISM_PATH_LIBRARY . '/libs');
 JLoader::registerNamespace('RandomLib', PRISM_PATH_LIBRARY . '/libs');
 JLoader::registerNamespace('SecurityLib', PRISM_PATH_LIBRARY . '/libs');
-JLoader::registerNamespace('AdamPaterson', PRISM_PATH_LIBRARY . '/libs');
-JLoader::registerNamespace('Carbon', PRISM_PATH_LIBRARY . '/libs');
 
 // Register some helpers.
 JHtml::addIncludePath(PRISM_PATH_LIBRARY .'/ui/helpers');